pyproject.toml
[tool.poetry]
name = "learning-fastapi"
version = "0.1.0"
description = ""
authors = ["David-Roklem <roklem@mail.ru>"]
readme = "README.md"
packages = [{include = "learning_fastapi"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.103.2"}
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
